clean-backend/requirements.txt
gunicorn==21.2.0
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
flask-bcrypt==1.0.1
python-dotenv==1.0.0
sendgrid==6.10.0
sqlalchemy==2.0.23
requests==2.31.0
